DR2-1255 Sort files based on sortOrder field.
There is now a sortOrder field in Dynamo. Files inside an asset need to
be ordered by this field before being added to the XIP.

Preservica decides the type of the asset by the first item in the
ContentObjects list and at the moment, that's json and we'd like it to
be docx.

There's a couple of other bits in here. The checksum field in Dynamo is
now called checksum_sha256 and there's some slight re-ordering of the
XIP.